Fix running litr in sub-folders failing with dir option
If the `dir` option was defined for a command and Litr was run from inside a sud-directory the execution failed as the directory could not be found. This was due to Litr setting a relativ path. Now the context for executing a command if `dir` is defined is an absolute path relative to the config file.

Closes #56