Fix wrongly ordered configuration file properties
So far commands and parameters read from the TOML configuration file were ordered alphabetically instead of in order of definition. This is now fixed and commands and parameters are read and processed in order of definition. This also means that they will now appear in order when calling `litr --help`.

Closes #65